+++ src/components/MyAccount/Settings/AccountSettingsTab.test.tsx
@@ -1,4 +1,4 @@
-import { patron } from "../../../../__test__/fixtures/myAccountFixtures"
+import { patron } from "../../../../__test__/fixtures/rawSierraAccountData"
 import AccountSettingsTab from "./AccountSettingsTab"
 import MyAccount from "../../../models/MyAccount"
 import { render, screen } from "../../../utils/testUtils"
